extract .js and .scss files from .vue in aeAddressInput
