+++ src/components/aeAddressInput/aeAddressInput.scss
@@ -1,0 +1,7 @@
+.ae-address-input.ae-textarea {
+  resize: none;
+  text-align: center;
+  min-height: 0;
+  height: 90px;
+  line-height: 29px;
+}
